even more linting stuff

+++ pkg/microservice/aslan/core/workflow/service/workflow/job/job_istio_release.go
@@ -185,7 +185,7 @@ func (j *IstioReleaseJob) LintJob() error {
 
 type lintIstioReleaseJob struct {
 	jobName string
-	weight  int
+	weight  int64
 }
 
 func lintFirstIstioReleaseJob(jobName string, stages []*commonmodels.WorkflowStage) error {
